requirements.txt
fastapi == 0.126.0
uvicorn == 0.44.0
sqlalchemy == 2.0.23
alembic == 1.18.4
python-multipart == 0.0.22
python-jose[cryptography] == 3.5.0
passlib == 1.7.4
pytest == 9.0.3
httpx == 0.28.1
psycopg2-binary==2.9.9